pyproject.toml
[tool.poetry]
name = "portsec-manager"
version = "0.0.1"
description = "Port-security manager"
authors = ["Gennadiy Tarada <gennadiy@tarada.xyz>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
nornir-scrapli = "^2021.7.30"
pysnmp = "^4.4.12"
rich = "^10.10.0"
requests = "^2.26.0"
netaddr = "^0.8.0"
gunicorn = "^20.1.0"
tabulate = "^0.8.9"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
mypy = "^0.910"
pytest = "^6.2.5"
flake8 = "^3.9.2"
yamllint = "^1.26.3"
pylint = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
